feature (dependent tasks): complete tasks serially
When the master server receives a worker-task-completed message it now
checks if all of the tasks have been completed.
If this is so, a complete-task message is sent to all workers for all
dependent tasks (this metadata is stored in the node recipes).
This allows for entire systems to be bounded and completed.