fix (tasks): add recipes through generated function
In the past, some OO magic was used to construct a method that would add
recipes to a master server.
However, this proved to have issues when there were multiple systems
declared within the same image.
Instead, a new function is built that allows the recipes to be easily
added to the master server.

ISSUE: #99